GovClerkMinutes-server/requirements.txt
urllib3==1.26.6
fastapi
uvicorn
uvloop
gunicorn
httptools
openai
pydub
python-dotenv
boto3
requests